Fixed check win condition/Add message for winner and invalid move
Fixed the check win condition by changing the order of operations.
Originally I had it in events.js when it should be checking right
after the board is updated. Also added messaging to tell players
who won and also if they clicked on an cell that already has a move.